+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 - Deprecated the usage of `renderInstance` (#465)
 - Removed support for IE11 (#467)
 - Removed bundlers (except default) from Piral monorepo
+- Added new bundler `piral-cli-parcel2` (#436)
+- Added new bundler `piral-cli-rollup` (#435)
 - Added new prop `meta` to registered pages
 - Added pre-bundled minified versions to each plugin package
 - Added support for *debug-meta.json* when using `pilet debug` (#462)